feat: added map location tagging feature
